requirements.txt
# requirements.txt
# Core RAG Framework
langchain
langchain-community

# Data Processing
pypdf

# Vector Database & Hybrid Search
chromadb==0.4.18
pydantic==1.10.13  # Required dependency version for Chroma

# Embedding Models & Hybrid Search (Sparse)
sentence-transformers
transformers
torch
rank_bm25          # For Sparse Retrieval (BM25)

# Reranking (Cross-Encoder)
torch
sentence-transformers

# Backend API
fastapi
uvicorn

# For saving the environment configuration (optional but good practice)
python-dotenv`